refactor: split examples utils into modular package structure
